cleanUp code (removed useless annotations and method's parameters) added tests corrected logic in methods

+++ src/main/java/pl/connectis/electronicswebshop/order/OrderService.java
@@ -1,16 +1,18 @@
 package pl.connectis.electronicswebshop.order;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.connectis.electronicswebshop.products.Product;
 import pl.connectis.electronicswebshop.service.OrderServiceInterface;
 
 import java.security.Principal;
-import java.util.Iterator;
+import java.util.Collection;
 import java.util.Optional;
 
 @Service
 @Transactional
+@Slf4j
 public class OrderService implements OrderServiceInterface {
 
     private final OrderRepository orderRepository;
@@ -23,11 +25,7 @@ public class OrderService implements OrderServiceInterface {
 
     @Override
     public Order addOrder() {
-        Order order = new Order();
-        order.setAddedBy("Anonymous");
-        order.setOrderStatus(OrderStatus.OPEN);
-        orderRepository.save(order);
-        return order;
+        return addOrder("Anonymous");
     }
 
     public Order addOrder(String username) {
@@ -36,15 +34,6 @@ public class OrderService implements OrderServiceInterface {
         order.setOrderStatus(OrderStatus.OPEN);
         orderRepository.save(order);
         return order;
-    }
-
-    @Override
-    public Optional<Order> findById(Long id) {
-        return orderRepository.findById(id);
-    }
-
-    public Iterable<Order> findAllByAddedBy(String addedBy) {
-        return orderRepository.findAllByAddedBy(addedBy);
     }
 
     @Override
@@ -57,10 +46,16 @@ public class OrderService implements OrderServiceInterface {
         return orderRepository.findById(lastOpenOrder.getId()).get().getProducts();
     }
 
-    public void deleteOrder(Principal principal) {
+    public boolean deleteOrder(Principal principal) {
+
         String username = ((principal == null) ? "Anonymous" : principal.getName());
         Order order = orderRepository.findByAddedByAndOrderStatus(username, OrderStatus.OPEN);
-        orderRepository.delete(order);
+        if (order != null) {
+            orderRepository.delete(order);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public Iterable<Order> getAllOrders() {
@@ -78,36 +73,43 @@ public class OrderService implements OrderServiceInterface {
             if (products.getProduct().equals(product)) {
                 orderLine = products;
                 orderLine.setQuantity(orderLine.getQuantity() + quantity);
-                product.setStock(product.getStock() - quantity);
             }
         }
         if (orderLine == null) {
             orderLine = new OrderLine(lastOpenOrder, product, quantity);
-            product.setStock(product.getStock() - quantity);
             lastOpenOrder.getProducts().add(orderLine);
         }
+        product.setStock(product.getStock() - quantity);
         saveOrder(lastOpenOrder);
         return true;
     }
 
-    public Order removeLine(long orderId, long productId, int quantity) {
+    public Order removeProduct(long orderId, long productId, int quantity) {
         Optional<Order> foundOrder = orderRepository.findById(orderId);
-        if (foundOrder.isPresent()) {
-            for (Iterator<OrderLine> iterator = foundOrder.get().getProducts().iterator(); iterator.hasNext(); ) {
 
-                OrderLine orderLine = iterator.next();
-                Product product = orderLine.getProduct();
-                if (orderLine.getProduct().getId().equals(productId)) {
-                    if (orderLine.getQuantity() - quantity <= 0) {
-                        product.setStock(orderLine.getQuantity() + quantity);
-                        iterator.remove();
-                        orderLineRepository.delete(orderLine);
-                    } else {
-                        product.setStock(orderLine.getQuantity() + quantity);
-                        orderLine.setQuantity(orderLine.getQuantity() - quantity);
-                    }
+
+        if (foundOrder.isPresent()) {
+            Collection<OrderLine> orderLinesCollection = foundOrder.get().getProducts();
+            OrderLine orderLine = null;
+            for (OrderLine currentOrderLine : orderLinesCollection) {
+                if (currentOrderLine.getProduct().getId().equals(productId)) {
+                    orderLine = currentOrderLine;
+                    break;
                 }
             }
+            if (orderLine != null) {
+                Product product = orderLine.getProduct();
+                if (orderLine.getQuantity() - quantity <= 0) {
+                    product.setStock(product.getStock() + orderLine.getQuantity());
+                    orderLinesCollection.remove(orderLine);
+                    orderLineRepository.delete(orderLine);
+                } else {
+                    product.setStock(product.getStock() + quantity);
+                    orderLine.setQuantity(orderLine.getQuantity() - quantity);
+
+                }
+            }
+
             return foundOrder.get();
         }
         return null;
